feat: diskon mendukung desimal (12.5% / 12.55%)
DISCOUNT_PERCENT berubah INT jadi DECIMAL(5,2) via Script0012; tipe
decimal? di MasterProduct, ProductDto, CreateProductRequest,
UpdateProductRequest; [Range(0,100)] tetap. Script data-bilangan
diperluas: mode desimal (flag data-bilangan-desimal) mengizinkan satu
pemisah '.' atau ',' (dinormalisasi ke '.'), max 2 angka desimal, clamp
100 tetap; sekaligus binding semua input data-bilangan (sebelumnya
querySelector hanya mengikat input pertama) sehingga input Stok ikut
tersanitasi. Input diskon pindah ke inputmode=decimal.

+++ cobaproject/Dtos/ProductDto.cs
@@ -10,7 +10,7 @@ public class ProductDto
     public string? Image { get; set; }
     public decimal? RatingRate { get; set; }
     public int? RatingCount { get; set; }
-    public int? DiscountPercent { get; set; }
+    public decimal? DiscountPercent { get; set; }
     public int Stock { get; set; }
     public bool IsActive { get; set; }
     public DateTime CreatedAt { get; set; }
