+++ cobaproject/Models/MasterProduct.cs
@@ -11,7 +11,7 @@ public class MasterProduct
     public string? Image { get; set; }
     public decimal? RatingRate { get; set; }
     public int? RatingCount { get; set; }
-    public int? DiscountPercent { get; set; }
+    public decimal? DiscountPercent { get; set; }
     public int Stock { get; set; }
 
     public bool IsActive { get; set; }
